- Main method added - IndexTable rewritten without STL vectors and with   DynamicArray

+++ src/module1/QueryScore.h
@@ -11,9 +11,10 @@
 #include <iostream>
 #include <cstring>
 #include <algorithm>
+#include <limits.h>
 
 typedef struct {
-    short seqId;
+    int seqId;
     short prefScore;
 } hit_t;
